--- public/app/percona/backup/components/BackupInventory/BackupCreation/BackupCreation.types.ts
@@ -1,3 +1,0 @@
-export interface BackupCreationProps {
-  date: number;
-}
