--- public/app/percona/backup/components/BackupInventory/BackupCreation/index.ts
@@ -1,1 +1,0 @@
-export * from './BackupCreation';
