+++ public/app/percona/backup/components/DetailedDate/DetailedDate.tsx
@@ -1,18 +1,18 @@
 import React, { FC } from 'react';
 import { useStyles } from '@grafana/ui';
 import moment from 'moment/moment';
-import { BackupCreationProps } from './BackupCreation.types';
-import { getStyles } from './BackupCreation.styles';
-import { DAY_FORMAT, HOUR_FORMAT } from './BackupCreation.constants';
+import { DetailedDateProps } from './DetailedDate.types';
+import { getStyles } from './DetailedDate.styles';
+import { DAY_FORMAT, HOUR_FORMAT } from './DetailedDate.constants';
 
-export const BackupCreation: FC<BackupCreationProps> = ({ date }) => {
+export const DetailedDate: FC<DetailedDateProps> = ({ date }) => {
   const styles = useStyles(getStyles);
   const momentObj = moment(date);
   const dayTime = momentObj.format(DAY_FORMAT);
   const hourTime = momentObj.format(HOUR_FORMAT);
 
   return (
-    <div data-qa="backup-creation">
+    <div data-qa="detailed-date">
       <span>{dayTime}</span>
       <span className={styles.hourWrapper}>{hourTime}</span>
     </div>
