+++ public/app/percona/backup/components/DetailedDate/DetailedDate.types.ts
@@ -1,0 +1,3 @@
+export interface DetailedDateProps {
+  date: number;
+}
