+++ public/app/percona/backup/components/DetailedDate/index.ts
@@ -1,0 +1,1 @@
+export * from './DetailedDate';
